修改deductBalance的类型为double

+++ ch-user-api/src/main/java/com/ai/ch/user/api/defaultlog/params/DefaultLogVo.java
@@ -39,7 +39,7 @@ public class DefaultLogVo extends BaseInfo{
 	/**
 	 * 扣款金额
 	 */
-	private Long deductBalance;
+	private double deductBalance;
 
 	/**
 	 * 操作员
@@ -83,11 +83,13 @@ public class DefaultLogVo extends BaseInfo{
 		this.defaultReason = defaultReason;
 	}
 
-	public Long getDeductBalance() {
+	
+
+	public double getDeductBalance() {
 		return deductBalance;
 	}
 
-	public void setDeductBalance(Long deductBalance) {
+	public void setDeductBalance(double deductBalance) {
 		this.deductBalance = deductBalance;
 	}
